requirements.txt
pydantic-ai>=0.0.48
pydantic>=2.0.0
rich>=13.0.0 
logfire>=0.1.0
python-dotenv>=1.0.0 